requirements.txt
streamlit
groq
python-dotenv
PyPDF2
